+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 - `highlight` property removed from `CodeBox` component
 - `Review` logo now can be customized, wihout choosing a predefined type (`logo` property now accept a standard JSX.Element)
 - UI re-designed for all components
+- New `Password` component, a modern password field
 
 <br>